Move logic for uploading manuscript out of the resolver and into the manuscripts controller. Also add pubsub and sciencbeamApi services as dependencies of the Manuscriptm controller to allow for easier testing in future
